add separate redirect config
